Extract taxonomy view into own component
